+++ app/static/assets/js/app/legacy-actions-customer-subscription.js
@@ -85,9 +85,6 @@ export function registerCustomerSubscriptionActions(actionRouter) {
         'complete-winback': () => {
             callLegacy('completeWinback');
         },
-        'select-customer': (payload) => {
-            callLegacy('selectCustomer', payload.customerId);
-        },
         'identify-caller-as-customer': (payload) => {
             callLegacy('identifyCallerAsCustomer', payload.customerId);
         },
@@ -102,12 +99,6 @@ export function registerCustomerSubscriptionActions(actionRouter) {
         },
         'revert-restitution': (payload) => {
             callLegacy('revertRestitution', payload.subId);
-        },
-        'change-contact-history-page': (payload) => {
-            callLegacy('changeContactHistoryPage', payload.newPage);
-        },
-        'toggle-timeline-item': (payload) => {
-            callLegacy('toggleTimelineItem', payload.entryDomId);
         },
         'select-offer': (payload, context) => {
             callLegacy('selectOffer', payload.offerId, payload.title, payload.description, context.event);